+++ src/emulator/cpu/csr.rs
@@ -14,7 +14,7 @@ pub struct ControlAndStatusRegister {
 }
 
 impl ControlAndStatusRegister {
-    pub fn csrrw(&mut self, address: u32, value: u32) -> u32 {
+    pub fn csrrw(&mut self, address: u64, value: u64) -> u64 {
         if self.ucsr.contains(address) {
             return self.ucsr.csrrw(address, value);
         }
@@ -27,7 +27,7 @@ impl ControlAndStatusRegister {
         panic!("todo: address not found.");
     }
 
-    pub fn csrrs(&mut self, address: u32, value: u32) -> u32 {
+    pub fn csrrs(&mut self, address: u64, value: u64) -> u64 {
         if self.ucsr.contains(address) {
             return self.ucsr.csrrs(address, value);
         }
@@ -40,7 +40,7 @@ impl ControlAndStatusRegister {
         panic!("todo: address not found.");
     }
 
-    pub fn csrrc(&mut self, address: u32, value: u32) -> u32 {
+    pub fn csrrc(&mut self, address: u64, value: u64) -> u64 {
         if self.ucsr.contains(address) {
             return self.ucsr.csrrc(address, value);
         }
@@ -55,8 +55,8 @@ impl ControlAndStatusRegister {
 }
 
 pub trait Csr {
-    fn contains(&self, address: u32) -> bool;
-    fn csrrw(&mut self, address: u32, value: u32) -> u32;
-    fn csrrs(&mut self, address: u32, value: u32) -> u32;
-    fn csrrc(&mut self, address: u32, value: u32) -> u32;
+    fn contains(&self, address: u64) -> bool;
+    fn csrrw(&mut self, address: u64, value: u64) -> u64;
+    fn csrrs(&mut self, address: u64, value: u64) -> u64;
+    fn csrrc(&mut self, address: u64, value: u64) -> u64;
 }
