+++ src/emulator/cpu/pc.rs
@@ -1,7 +1,7 @@
 use crate::emulator::memory::MEMORY_BASE_ADDRESS;
 
 pub struct ProgramCounter {
-    pc: u32,
+    pc: u64,
 }
 
 impl Default for ProgramCounter {
@@ -13,7 +13,7 @@ impl Default for ProgramCounter {
 }
 
 impl ProgramCounter {
-    pub fn read(&self) -> u32 {
+    pub fn read(&self) -> u64 {
         self.pc
     }
 
@@ -21,12 +21,12 @@ impl ProgramCounter {
         self.pc += 4;
     }
 
-    pub fn jump(&mut self, address: u32) {
+    pub fn jump(&mut self, address: u64) {
         self.pc = address;
     }
 
-    pub fn jumpr(&mut self, relative_address: i32) {
-        self.pc = (self.pc as i64 + relative_address as i64) as u32;
+    pub fn jumpr(&mut self, relative_address: i64) {
+        self.pc = (self.pc as i64 + relative_address) as u64;
     }
 
     #[allow(dead_code)]
