+++ src/emulator/memory.rs
@@ -1,7 +1,7 @@
 use crate::emulator::bus::Xlen;
 
-pub const MEMORY_SIZE: u32 = 1024 * 1024 * 1024;
-pub const MEMORY_BASE_ADDRESS: u32 = 0x8000_0000;
+pub const MEMORY_SIZE: u64 = 1024 * 1024 * 1024;
+pub const MEMORY_BASE_ADDRESS: u64 = 0x8000_0000;
 
 pub struct Memory {
     pub memory: Vec<u8>,
@@ -16,17 +16,17 @@ impl Default for Memory {
 }
 
 impl Memory {
-    pub fn size(&self) -> u32 {
+    pub fn size(&self) -> u64 {
         MEMORY_BASE_ADDRESS + MEMORY_SIZE
     }
 
-    pub fn load(&self, address: u32, xlen: Xlen) -> u32 {
+    pub fn load(&self, address: u64, xlen: Xlen) -> u64 {
         (0..xlen as usize).fold(0, |acc, i| {
-            acc | (self.memory[(address - MEMORY_BASE_ADDRESS) as usize + i] as u32) << (8 * i)
+            acc | (self.memory[(address - MEMORY_BASE_ADDRESS) as usize + i] as u64) << (8 * i)
         })
     }
 
-    pub fn store(&mut self, address: u32, value: u32, xlen: Xlen) {
+    pub fn store(&mut self, address: u64, value: u64, xlen: Xlen) {
         for i in 0..xlen as usize {
             self.memory[(address - MEMORY_BASE_ADDRESS) as usize + i] = (value >> (i * 8)) as u8;
         }
